+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -1,8 +1,8 @@
-package bowling.domain;
+package bowling.domain.frame;
 
-import bowling.score.FinalScores;
-import bowling.score.NormalScores;
-import bowling.score.Scores;
+import bowling.domain.score.FinalScores;
+import bowling.domain.score.NormalScores;
+import bowling.domain.score.Scores;
 
 public class NormalFrame extends Frame {
