+++ src/main/java/bowling/domain/score/NormalScores.java
@@ -1,4 +1,4 @@
-package bowling.score;
+package bowling.domain.score;
 
 import java.util.Arrays;
 import java.util.List;
