+++ src/main/java/bowling/domain/score/Score.java
@@ -1,4 +1,4 @@
-package bowling.score;
+package bowling.domain.score;
 
 import bowling.exception.ExceptionMessage;
