+++ src/test/java/bowling/domain/score/ScoresTest.java
@@ -1,4 +1,4 @@
-package bowling.score;
+package bowling.domain.score;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
